requirements.txt
torch
torchvision
streamlit
Pillow
numpy
scikit-learn
google-generativeai
